requirements.txt
browser-use
openai
langchain-openai
python-dotenv
asyncio
beautifulsoup4
streamlit
flask
requests
pydantic
